Major refactoring to make things simpler.
