+++ discovery_engine_core/web-api2/Cargo.toml
@@ -2,7 +2,6 @@
 name = "web-api2"
 version = "0.1.0"
 edition = "2021"
-
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
